fix: compilation errors - jakarta validation, PageResult type, swagger dependency
- Replace javax.validation with jakarta.validation (Spring Boot 3.x)
- Fix PageResult.of() type casting (int to Long)
- Fix List to Long[] conversion
- Replace knife4j with springdoc-openapi
- Fix missing imports

+++ backend/oa-platform-admin/src/main/java/com/oa/system/service/UserService.java
@@ -51,7 +51,7 @@ public class UserService {
         for (SysUser user : list) {
             voList.add(toUserListVo(user));
         }
-        return PageResult.of(pageInfo.getTotal(), voList, pageInfo.getPageNum(), pageInfo.getPageSize());
+        return PageResult.of((long)pageInfo.getTotal(), voList, (long)pageInfo.getPageNum(), (long)pageInfo.getPageSize());
     }
 
     /**
